refactor: move everything into LandingPad class
i have a problem tbh
