+++ src/main/java/loqor/ait/tardis/Tardis.java
@@ -11,7 +11,7 @@ import loqor.ait.tardis.base.Initializable;
 import loqor.ait.tardis.base.TardisComponent;
 import loqor.ait.tardis.control.sequences.SequenceHandler;
 import loqor.ait.tardis.data.*;
-import loqor.ait.tardis.data.landing.LandingPadHandler;
+import loqor.ait.tardis.data.landing.LandingPad;
 import loqor.ait.tardis.data.loyalty.LoyaltyHandler;
 import loqor.ait.tardis.data.permissions.PermissionHandler;
 import loqor.ait.tardis.data.travel.TravelHandler;
@@ -219,7 +219,7 @@ public abstract class Tardis extends Initializable<TardisComponent.InitContext> 
         return this.handler(TardisComponent.Id.CLOAK);
     }
 
-    public LandingPadHandler landingPad() {
+    public LandingPad.Handler landingPad() {
         return this.handler(TardisComponent.Id.LANDING_PAD);
     }
 }
